QThread: Work around potential crash during objectName access
We access objectName from a different thread than the thread which owns
it. This is inherently racy, but was apparently not an issue in practice
before. With the new binding system, we however seem to encounter a case
where we do run into issues. This patch migitates it by bypassing the
binding system (using valueBypassingBindings).

Task-number: QTBUG-96718
Pick-to: 6.2 dev
Change-Id: I9d8f2354ae282893a32c8a23381d0056ceef7c40
Reviewed-by: Lars Knoll <lars.knoll@qt.io>